test: cleanup test_utils.rs
